* Flag carrier is shown on the map in WSG and EotS * .battleground startbg * Modified countdown before bg starts * Show message when wsg flag is picked up from the ground * Correct bg icon for Horde by Gaijin
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@288 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/ascent-world/SpellAuras.h
@@ -678,7 +678,7 @@ public:
 
 	bool HasMechanic(uint32 MechanicsType)
 	{
-		return (uint32)m_spellProto->MechanicsType == MechanicsType ||
+		return m_spellProto->MechanicsType == MechanicsType ||
 			(uint32)m_spellProto->EffectMechanic[0] == MechanicsType ||
 			(uint32)m_spellProto->EffectMechanic[1] == MechanicsType ||
 			(uint32)m_spellProto->EffectMechanic[2] == MechanicsType;
